refactor: player 패키지 이름 participant로 변경
