+++ src/main/java/blackjack/domain/participant/Player.java
@@ -1,4 +1,4 @@
-package blackjack.domain.player;
+package blackjack.domain.participant;
 
 import blackjack.domain.card.Hand;
