+++ src/main/java/blackjack/domain/participant/PlayerName.java
@@ -1,4 +1,4 @@
-package blackjack.domain.player;
+package blackjack.domain.participant;
 
 import java.util.Objects;
 
@@ -23,8 +23,12 @@ public class PlayerName {
 
     @Override
     public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
+        if (this == o) {
+            return true;
+        }
+        if (o == null || getClass() != o.getClass()) {
+            return false;
+        }
         PlayerName that = (PlayerName) o;
         return Objects.equals(value, that.value);
     }
