+++ src/main/java/blackjack/domain/participant/Players.java
@@ -1,4 +1,4 @@
-package blackjack.domain.player;
+package blackjack.domain.participant;
 
 import java.util.List;
