+++ src/test/java/blackjack/domain/participant/ParticipantsTest.java
@@ -1,4 +1,4 @@
-package blackjack.domain.player;
+package blackjack.domain.participant;
 
 import static org.assertj.core.api.Assertions.assertThat;
