+++ src/test/java/blackjack/domain/rule/JudgeTest.java
@@ -5,11 +5,11 @@ import blackjack.domain.card.Card;
 import blackjack.domain.card.CardRank;
 import blackjack.domain.card.CardSuit;
 import blackjack.domain.card.Hand;
-import blackjack.domain.player.Dealer;
-import blackjack.domain.player.Participants;
-import blackjack.domain.player.Player;
-import blackjack.domain.player.PlayerName;
-import blackjack.domain.player.Players;
+import blackjack.domain.participant.Dealer;
+import blackjack.domain.participant.Participants;
+import blackjack.domain.participant.Player;
+import blackjack.domain.participant.PlayerName;
+import blackjack.domain.participant.Players;
 import fixture.DealerFixture;
 import fixture.HandFixture;
 import fixture.PlayerFixture;
