+++ src/test/java/fixture/DealerFixture.java
@@ -4,7 +4,7 @@ import blackjack.domain.card.Card;
 import blackjack.domain.card.CardRank;
 import blackjack.domain.card.CardSuit;
 import blackjack.domain.card.Hand;
-import blackjack.domain.player.Dealer;
+import blackjack.domain.participant.Dealer;
 import java.util.ArrayList;
 import java.util.List;
